second commit

+++ myshop/shop/products/forms.py
@@ -1,11 +1,11 @@
 from flask_wtf.file import FileAllowed, FileField, FileRequired, ValidationError
-from wtforms import Form, IntegerField, StringField, BooleanField, TextAreaField, validators
+from wtforms import Form, IntegerField, StringField, BooleanField, TextAreaField, validators, DecimalField
 
 
 
 class Addproducts(Form):
     name = StringField('Name', [validators.DataRequired()])
-    price = IntegerField('Price', [validators.DataRequired()])
+    price = DecimalField('Price', [validators.DataRequired()])
     discount = IntegerField('Discount', default=0)
     stock = IntegerField('Stock', [validators.DataRequired()])
     description = TextAreaField('Description', [validators.DataRequired()])
